[feat+perf] Optimize performance by changing temperature and coordinate data types from Double to Float; implement offline-first weather information for the displayed location; rewrite the initial control flow in MainViewModel and MainActivity

+++ app/src/main/java/com/example/weatherjourney/app/navigation/WtnNavGraph.kt
@@ -54,7 +54,7 @@ fun WtnNavGraph(
                 val countryCode = it.getString(COUNTRY_CODE_ARG) ?: ""
 
                 WeatherInfoScreen(
-                    coordinate = Coordinate(latitude.toDouble(), longitude.toDouble()),
+                    coordinate = Coordinate(latitude, longitude),
                     snackbarHostState = snackbarHostState,
                     navigationKey = navigationKey,
                     countryCode = countryCode,
